requirements/local.txt
-r ./base.txt
django-debug-toolbar==3.4.0
coverage<4.6
pytest-django<3.5
